Move the copy/paste cursor to its own file
